fix bad includes for SDL2
The pkg-config files all add /usr/include/SDL2 as an include directory,
and upstream expects headers to be included without a directory
component.

Using the directory component anyway, may sometimes work in the
exclusive event that SDL2 is installed directly to /usr/include and that
is the compiler's default search path, and thus the SDL2/ directory is
itself in the compiler's builtin default search path.

But this is merely a common coincidence, semantically wrong, and hardly
guaranteed. And it failed for me.