meson: use the gnome postinstall function instead of custom scripts
They do the same thing, but the builtin meson one is shorter and has
better logging.

Since this requires a relatively recent version of meson (only a year
old!) it may not be available by default everywhere. Although it is easy
to install a newer version, add a minimum version declaration so that
older versions of meson can report a more informative error when using
too old a version.

Also install it in CI. The latest version of Ubuntu is new enough, but
the github workflow isn't using the latest version of Ubuntu.