fix desync issues when moving before the beginning
capped the GoTo method to the starting position of the music

On reset, the individual midi tracks were set to the 0 position but the midi itself had a negative time (for the lead in). There were some tricks in the Update method to work around this, which otherwise would cause the tracks to start playing at the wrong time. These workarounds were removed and reset method of the Midi also sets the tracks to the same negative time. Both times should now always be in sync

closes #73